Table of content that works  :)
Table of content is the crucial feature for me. So after digging in the
wkhtmltopdf source, I understood the process of creating ToC.
This commit contains "wkhtmltopdf 0.12.2", some minor cleanups and
whole new TableOfContentSettings. It behaves almost like in original
"wkhtmltopdf" - there's an object of type TableOfContentSettings which
triggers table of content creation. If there's no Xslt Transformation
given, Pechkin is creating a new one like in "wkhtmltopdf".
Test application produce pdf with table of content.